Add main package
This commit adds the main.go file and prints "Hello world!" text. To
try it do this:
go run main.go

Pay attention to the commit message I've made. There is a title which
should be less than 50 charachters, and there is a body that you can
explain why this commit is made.